requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
trafilatura>=1.12.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pytest>=8.0.0
